# apis/api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: LunarDB API
  version: 1.0.0
  description: API for interacting with LunarDB

paths:
  /api/v1/execute:
    post:
      summary: Execute a LunarDB command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
  /api/v1/set:
    post:
      summary: Set a key-value pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                value:
                  type: string
                ttl:
                  type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
  /api/v1/get/{key}:
    get:
      summary: Get the value for a key
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
  /api/v1/keys:
    get:
      summary: List all keys
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      type: string

components:
  schemas: {}